# perf: Remove usage of nested lmdb transaction