Spotted this in the flamegraph.

## Test method

Tested using data and config of `taxi-objectstore` in https://github.com/getdozer/dozer-samples-internal/commit/e4b719153a9d82f5fba55bb9637d44cd946f6304. Run following command:

```bash
rm -r .dozer; time ~/dozer/target/release/dozer
```

When `trips_cache` count reaches 200,000, press Ctrl + C and wait for dozer to quit.

## Result

With code in https://github.com/getdozer/dozer/commit/0c549a28f39c3f440b4768412fff7c616b7eaa5e:

`~/dozer/target/release/dozer  28.25s user 23.84s system 101% cpu 51.526 total`

With code in https://github.com/chubei/dozer/commit/ac46db09aca5b0751b07aaf69bb28acda872ee48:

`~/dozer/target/release/dozer  13.62s user 10.19s system 115% cpu 20.578 total`

Saves 60%.